notes/requirement.txt
ansicon==1.89.0
asgiref==3.5.2
asttokens==2.2.1
babelfish==0.6.0
backcall==0.2.0
blessed==1.19.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
codefind==0.1.3
colorama==0.4.6
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.5
Django==4.1.7
django-cors-headers==3.13.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.7.1
fonttools==4.38.0
future==0.18.2
idna==3.4
ipykernel==6.17.1
ipython==8.7.0
jedi==0.18.2
jinxed==1.2.0
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jurigged==0.5.3
kiwisolver==1.4.4
MapProxy==1.15.1
Markdown==3.4.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mysql-connector==2.2.9
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.23.4
opencv-python==4.6.0.66
openpyxl==3.0.10
ovld==0.3.2
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pkg==0.2
platformdirs==2.5.2
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg==3.1.8
psycopg2-binary==2.9.5
pure-eval==0.2.2
pycparser==2.21
pydub==0.25.1
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pypng==0.20220715.0
pyproj==3.4.1
PyQRCode==1.2.1
python-dateutil==2.8.2
pytz==2022.6
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
qrcode==7.4.2
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.2
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.6.0
typing_extensions==4.5.0
tzdata==2022.1
urllib3==1.26.12
virtualenv==20.15.1
vonage==3.1.0
watchdog==2.1.9
wcwidth==0.2.5
wrapt==1.14.1
